BAU: Move common variables to a separate file
These test data can be re-used beyond the service tests, so moving these to somewhere common
