+++ src/components/reset-password/tests/reset-password-service.test.ts
@@ -5,7 +5,6 @@ import { resetPasswordService } from "../reset-password-service";
 import { Http } from "../../../utils/http";
 import {
   checkApiCallMadeWithExpectedBodyAndHeaders,
-  commonVariables,
   expectedHeadersFromCommonVarsWithSecurityHeaders,
   requestHeadersWithIpAndAuditEncoded,
   resetApiKeyAndBaseUrlEnvVars,
@@ -17,6 +16,7 @@ import {
   PATH_NAMES,
 } from "../../../app.constants";
 import { createMockRequest } from "../../../../test/helpers/mock-request-helper";
+import { commonVariables } from "../../../../test/helpers/common-test-variables";
 
 describe("reset password service", () => {
   const httpInstance = new Http();
