+++ test/helpers/common-test-variables.ts
@@ -1,0 +1,10 @@
+export const commonVariables = {
+  email: "joe.bloggs@example.com",
+  sessionId: "some-session-id",
+  diPersistentSessionId: "some-persistent-session-id",
+  clientSessionId: "some-client-session-id",
+  ip: "123.123.123.123",
+  apiKey: "apiKey",
+  auditEncodedString:
+    "R21vLmd3QilNKHJsaGkvTFxhZDZrKF44SStoLFsieG0oSUY3aEhWRVtOMFRNMVw1dyInKzB8OVV5N09hOi8kLmlLcWJjJGQiK1NPUEJPPHBrYWJHP358NDg2ZDVc",
+};
