+++ test/helpers/service-test-helper.ts
@@ -2,17 +2,7 @@ import { ApiResponseResult } from "../../src/types";
 import { SinonStub } from "sinon";
 import { expect } from "chai";
 import { sinon } from "../utils/test-utils";
-
-export const commonVariables = {
-  email: "joe.bloggs@example.com",
-  sessionId: "some-session-id",
-  diPersistentSessionId: "some-persistent-session-id",
-  clientSessionId: "some-client-session-id",
-  ip: "123.123.123.123",
-  apiKey: "apiKey",
-  auditEncodedString:
-    "R21vLmd3QilNKHJsaGkvTFxhZDZrKF44SStoLFsieG0oSUY3aEhWRVtOMFRNMVw1dyInKzB8OVV5N09hOi8kLmlLcWJjJGQiK1NPUEJPPHBrYWJHP358NDg2ZDVc",
-};
+import { commonVariables } from "./common-test-variables";
 
 export const expectedHeadersFromCommonVarsWithSecurityHeaders = {
   "X-API-Key": commonVariables.apiKey,
